+++ src/interval.h
@@ -3,24 +3,24 @@
 #include "common.h"
 class interval {
 public:
-    const double min, max;
+    const float min, max;
 
     constexpr interval() : min(-infinity), max(infinity) {}
-    constexpr interval(double min, double max) : min(min), max(max) {}
+    constexpr interval(float min, float max) : min(min), max(max) {}
 
-    double size() const {
+    float size() const {
         return max - min;
     }
 
-    bool contains(double x) const {
+    bool contains(float x) const {
         return x >= min && x <= max;
     }
 
-    bool surrounds(double x) const {
+    bool surrounds(float x) const {
         return x > min && x < max;
     }
 
-    double clamp(double x) const {
+    float clamp(float x) const {
         if (x < min) return min;
         if (x > max) return max;
         return x;
